server/requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
lxml==5.4.0
packaging==25.0
PyJWT==2.9.0
python-docx==1.2.0
requests==2.32.4
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.4.0
whitenoise==6.9.0
